GetUserPosts service function created
